optimize main menu behaviour: add ResponsiveNavigation.js, split desktop and mobile menu stylesheet
the ResponsiveNavigation class changes the class of its target element (and child elements) and adds/removes its modal functionality, based on the current body width.
This ensures the menu element only has a Modal instance attached to it when really necessary, fixing some tabbing/accessibility issues.
Changing the classes makes it more comfortable to style the desktop and mobile view of the menu